add kernel api tests
